# openapi/wms-api.yaml
openapi: 3.1.0
info:
  title: WMS ShowCase API
  description: 倉庫管理システム（WMS）ShowCase のREST API定義
  version: 0.1.0
  contact:
    name: WMS ShowCase Project
servers:
  - url: http://localhost:8080
    description: ローカル開発環境
  - url: https://{environment}.azurecontainerapps.io
    description: Azure Container Apps
    variables:
      environment:
        default: wms-showcase
...
paths:
...
  /api/v1/master/areas:
    get:
      operationId: listAreas
      summary: エリア一覧取得
      description: エリアマスタの一覧を取得する
      tags: [master-area]
      parameters:
        - name: warehouseId
          in: query
          schema:
            type: integer
            format: int64
            minimum: 1
          description: 倉庫IDで絞り込み
        - name: buildingId
          in: query
          schema:
            type: integer
            format: int64
            minimum: 1
          description: 棟IDで絞り込み
        - name: storageCondition
          in: query
          schema:
            $ref: '#/components/schemas/StorageCondition'
          description: 保管条件フィルタ
        - name: areaType
          in: query
          schema:
            $ref: '#/components/schemas/AreaType'
          description: エリア種別フィルタ
        - name: isActive
          in: query
          schema:
            type: boolean
          description: 有効/無効フィルタ
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: エリア一覧取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AreaPageResponse'
        '400':
          $ref: '#/components/responses/InvalidParameter'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: createArea
      summary: エリア登録
      description: エリアマスタに新規エリアを登録する
      tags: [master-area]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAreaRequest'
      responses:
        '201':
          description: エリア登録成功
          headers:
            Location:
              description: 作成されたリソースのURL
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AreaDetail'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: 指定棟が存在しない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: エリアコード重複
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                duplicateCode:
                  value:
                    errorCode: DUPLICATE_CODE
                    message: 同一棟内に同じエリアコードが既に存在します
  /api/v1/master/areas/{id}:
    get:
      operationId: getArea
      summary: エリア取得
      description: 指定IDのエリアマスタを1件取得する
      tags: [master-area]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: エリア取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AreaDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: エリアが存在しない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                notFound:
                  value:
                    errorCode: AREA_NOT_FOUND
                    message: 指定されたエリアが見つかりません
    put:
      operationId: updateArea
      summary: エリア更新
      description: 指定IDのエリアマスタを更新する。エリアコード・棟ID・エリア種別は変更不可
      tags: [master-area]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAreaRequest'
      responses:
        '200':
          description: エリア更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AreaUpdateResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: エリアが存在しない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          $ref: '#/components/responses/OptimisticLockConflict'
  /api/v1/master/areas/{id}/toggle-active:
    patch:
      operationId: toggleAreaActive
      summary: エリア無効化／有効化
      description: 指定IDのエリアの有効/無効を切り替える。配下にロケーションが存在するエリアは無効化不可
      tags: [master-area]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ToggleActiveRequest'
      responses:
        '200':
          description: 無効化/有効化成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AreaToggleResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: エリアが存在しない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          $ref: '#/components/responses/OptimisticLockConflict'
        '422':
          description: 配下にロケーションが存在するため無効化不可
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                hasChildren:
                  value:
                    errorCode: CANNOT_DEACTIVATE_HAS_CHILDREN
                    message: 配下にロケーションが存在するため無効化できません

  # --- Locations ---
...
components:
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: JWTアクセストークン（httpOnly Cookie）

  # ============================================================
  # Schemas
  # ============================================================
  schemas:
...
    ErrorResponse:
      type: object
      required: [errorCode, message]
      properties:
        errorCode:
          type: string
          description: エラーコード
          example: VALIDATION_ERROR
        message:
          type: string
          description: エラーメッセージ
          example: 入力内容に誤りがあります
        details:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
          description: フィールドレベルのバリデーションエラー詳細（VALIDATION_ERRORの場合のみ）
    FieldError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
          description: エラーが発生したフィールド名
          example: userCode
        message:
          type: string
          description: フィールドのエラーメッセージ
          example: 必須項目です
...
    StorageCondition:
      type: string
      enum: [AMBIENT, REFRIGERATED, FROZEN]
      description: 保管条件（常温・冷蔵・冷凍）
    AreaType:
      type: string
      enum: [STOCK, INBOUND, OUTBOUND, RETURN]
      description: エリア種別（在庫・入荷・出荷・返品）

    # --- Facility Master: Common ---
    ToggleActiveRequest:
      type: object
      required: [isActive, version]
      properties:
        isActive:
          type: boolean
          description: falseで無効化、trueで有効化
        version:
          type: integer
          description: 楽観的ロックバージョン

    # --- Facility Master: Warehouse ---
...
    CreateAreaRequest:
      type: object
      required: [buildingId, areaCode, areaName, storageCondition, areaType]
      properties:
        buildingId:
          type: integer
          format: int64
          description: 所属棟ID
        areaCode:
          type: string
          maxLength: 20
          pattern: '^[A-Za-z0-9\-]+$'
          description: エリアコード（棟内一意・変更不可）
          example: A01
        areaName:
          type: string
          maxLength: 200
          description: エリア名称
          example: A棟1階 常温エリア
        storageCondition:
          $ref: '#/components/schemas/StorageCondition'
        areaType:
          $ref: '#/components/schemas/AreaType'
    UpdateAreaRequest:
      type: object
      required: [areaName, storageCondition, version]
      properties:
        areaName:
          type: string
          maxLength: 200
          description: エリア名称
        storageCondition:
          $ref: '#/components/schemas/StorageCondition'
        version:
          type: integer
          description: 楽観的ロックバージョン
    AreaDetail:
      type: object
      required: [id, areaCode, areaName, warehouseId, warehouseCode, buildingId, buildingCode, storageCondition, areaType, isActive, version, createdAt, updatedAt]
      properties:
        id:
          type: integer
          format: int64
        areaCode:
          type: string
        areaName:
          type: string
        warehouseId:
          type: integer
          format: int64
        warehouseCode:
          type: string
        warehouseName:
          type: string
        buildingId:
          type: integer
          format: int64
        buildingCode:
          type: string
        buildingName:
          type: string
        storageCondition:
          $ref: '#/components/schemas/StorageCondition'
        areaType:
          $ref: '#/components/schemas/AreaType'
        isActive:
          type: boolean
        version:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AreaPageResponse:
      type: object
      required: [content, page, size, totalElements, totalPages]
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/AreaListItem'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
    AreaListItem:
      type: object
      required: [id, areaCode, areaName, warehouseId, warehouseCode, buildingId, buildingCode, storageCondition, areaType, isActive, version, createdAt, updatedAt]
      properties:
        id:
          type: integer
          format: int64
        areaCode:
          type: string
        areaName:
          type: string
        warehouseId:
          type: integer
          format: int64
        warehouseCode:
          type: string
        buildingId:
          type: integer
          format: int64
        buildingCode:
          type: string
        storageCondition:
          $ref: '#/components/schemas/StorageCondition'
        areaType:
          $ref: '#/components/schemas/AreaType'
        isActive:
          type: boolean
        version:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AreaUpdateResponse:
      type: object
      required: [id, areaCode, areaName, buildingId, storageCondition, areaType, isActive, version, updatedAt]
      properties:
        id:
          type: integer
          format: int64
        areaCode:
          type: string
        areaName:
          type: string
        buildingId:
          type: integer
          format: int64
        storageCondition:
          $ref: '#/components/schemas/StorageCondition'
        areaType:
          $ref: '#/components/schemas/AreaType'
        isActive:
          type: boolean
        version:
          type: integer
        updatedAt:
          type: string
          format: date-time
    AreaToggleResponse:
      type: object
      required: [id, areaCode, areaName, isActive, version, updatedAt]
      properties:
        id:
          type: integer
          format: int64
        areaCode:
          type: string
        areaName:
          type: string
        isActive:
          type: boolean
        version:
          type: integer
        updatedAt:
          type: string
          format: date-time

    # --- Facility Master: Location ---
...
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
      description: リソースID
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
      description: ページ番号（0始まり）
    SizeParam:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: ページサイズ
...
  responses:
    ValidationError:
      description: バリデーションエラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validationError:
              value:
                errorCode: VALIDATION_ERROR
                message: 入力内容に誤りがあります
                details:
                  - field: userCode
                    message: 必須項目です
    Unauthorized:
      description: 認証エラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unauthorized:
              value:
                errorCode: UNAUTHORIZED
                message: 認証が必要です
    Forbidden:
      description: 権限不足
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            forbidden:
              value:
                errorCode: FORBIDDEN
                message: この操作を実行する権限がありません
    InvalidParameter:
      description: クエリパラメータ不正
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalidParameter:
              value:
                errorCode: INVALID_PARAMETER
                message: クエリパラメータが不正です
    OptimisticLockConflict:
      description: 楽観的ロック競合
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            conflict:
              value:
                errorCode: OPTIMISTIC_LOCK_CONFLICT
                message: 他のユーザーによる更新が先行しました。最新データを取得してから再度お試しください
security:
  - cookieAuth: []